roles now can be more than one. also changed the types of id now are u64. fixed readm.me , etc...

+++ src/views/upload_profile_picture.rs
@@ -11,8 +11,8 @@ use crate::jwt::validate;
 
 #[derive(serde::Deserialize, serde::Serialize)]
 struct Task{
-    id: i32,
-    grade: i32,
+    id: u64,
+    grade: u64,
 }
